+++ src/Build5Nines.SharpVector.OpenAI/OpenAIMemoryVectorDatabaseBase.cs
@@ -251,13 +251,19 @@ public abstract class OpenAIMemoryVectorDatabaseBase<TId, TMetadata, TVectorStor
         return results;
     }
 
-        /// <summary>
+    [Obsolete("Use SerializeToBinaryStreamAsync Instead")]
+    public virtual async Task SerializeToJsonStreamAsync(Stream stream)
+    {
+        await SerializeToBinaryStreamAsync(stream);
+    }
+
+    /// <summary>
     /// Serializes the Vector Database to a JSON stream
     /// </summary>
     /// <param name="stream"></param>
     /// <returns></returns>
     /// <exception cref="InvalidOperationException"></exception>
-    public virtual async Task SerializeToJsonStreamAsync(Stream stream)
+    public virtual async Task SerializeToBinaryStreamAsync(Stream stream)
     {
         var streamVectorStore = new MemoryStream();
         var streamVocabularyStore = new MemoryStream();
@@ -296,16 +302,28 @@ public abstract class OpenAIMemoryVectorDatabaseBase<TId, TMetadata, TVectorStor
         await stream.FlushAsync();
     }
 
+    [Obsolete("Use SerializeToBinaryStream Instead")]
     public virtual void SerializeToJsonStream(Stream stream)
+    {
+        SerializeToBinaryStream(stream);
+    }
+
+    public virtual void SerializeToBinaryStream(Stream stream)
     {
         if (stream == null)
         {
             throw new ArgumentNullException(nameof(stream));
         }
-        SerializeToJsonStreamAsync(stream).Wait();
+        SerializeToBinaryStreamAsync(stream).Wait();
     }
 
+    [Obsolete("Use DeserializeFromBinaryStreamAsync Instead")]
     public virtual async Task DeserializeFromJsonStreamAsync(Stream stream)
+    {
+        await DeserializeFromBinaryStreamAsync(stream);
+    }
+
+    public virtual async Task DeserializeFromBinaryStreamAsync(Stream stream)
     {
         if (stream == null)
         {
@@ -374,13 +392,19 @@ public abstract class OpenAIMemoryVectorDatabaseBase<TId, TMetadata, TVectorStor
         }
     }
 
+    [Obsolete("Use DeserializeFromBinaryStream Instead")]
     public virtual void DeserializeFromJsonStream(Stream stream)
+    {
+        DeserializeFromBinaryStream(stream);
+    }
+    
+    public virtual void DeserializeFromBinaryStream(Stream stream)
     {
         if (stream == null)
         {
             throw new ArgumentNullException(nameof(stream));
         }
-        DeserializeFromJsonStreamAsync(stream).Wait();
+        DeserializeFromBinaryStreamAsync(stream).Wait();
     }
 
 }